refactor: 공통 사용 DTO common 패키지로 이동
