+++ src/main/java/kitchenpos/application/dto/common/ProductResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.dto;
+package kitchenpos.application.dto.common;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.math.BigDecimal;
