+++ src/main/java/kitchenpos/ui/dto/CreateMenuRequest.java
@@ -5,7 +5,7 @@ import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.application.dto.CreateMenuCommand;
-import kitchenpos.application.dto.MenuProductCommand;
+import kitchenpos.application.dto.common.MenuProductCommand;
 
 public class CreateMenuRequest {
