+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -9,8 +9,8 @@ import java.math.BigDecimal;
 import java.util.List;
 import kitchenpos.application.dto.CreateMenuCommand;
 import kitchenpos.application.dto.CreateMenuResponse;
-import kitchenpos.application.dto.MenuProductCommand;
 import kitchenpos.application.dto.SearchMenuResponse;
+import kitchenpos.application.dto.common.MenuProductCommand;
 import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuGroup;
 import kitchenpos.domain.MenuProduct;
